pyproject.toml
# see https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
# see https://docs.astral.sh/uv/

# ビルドシステムの指定をすると uv sync で自動でリポジトリをインストール
# see https://docs.astral.sh/uv/concepts/projects/config/#build-systems
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "video_converter"
requires-python = "==3.12.*"
version = "0.1.0"
dependencies = [
    # interfaces
    "fire",
    "pyinstaller",
    "tqdm",
    "tkinterdnd2",
    # libraries
    "psutil",
    "gevent",
    "ffmpeg-python",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.11",
    "sphinx>=9.1.0",
    "sphinx-mdinclude>=0.6.2",
    "sphinx-rtd-theme>=0.5.1",
]
